openvswitch-switch.postinst: Avoid getting redirected to systemctl.
When the openvswitch-switch package is installed, postinst script
is used to make a call to /etc/init.d/openvswitch-switch with
"restart" as the command. openvswitch-switch.init inturn would carefully
restart OVS by saving the openflow flows and replacing the kernel
module only when needed.

With recent versions of Ubuntu and Debian, any call of "restart"
no longer calls the "restart" of openvswitch-switch.init but instead
calls a runtime generated systemd file which inturn makes a "stop"
call followed by a "start" call. This would mean that we would neither
save flows or replace the kernel module.

This commit tries to solve this problem by getting rid of '#DEBHELPER'
and replacing "restart" with "ovsrestart" so that we skip systemd
generated files.

VMware-BZ: #1890631
Signed-off-by: Gurucharan Shetty <guru@ovn.org>