ovs-thread: Retry on read lock.
We find a bug on Ubuntu 18.04 system where the revalidator thread is
blocked for indefinitely long time on waiting for a rwlock_rdlock with
the following stack trace. After digging into the kernel source tree
we find that it may related to a recent reported futex missed wakeup
issue.
    * https://github.com/torvalds/linux/commit/b061c38bef43

This patch try to work around the missed wakeup by checking the read
lock every second.  We will use this patch to identify the root cause
of the revalidator thread block issue.

Signed-off-by: Yi-Hung Wei <yihung.wei@gmail.com>
Co-authored-by: Ansis Atteka <aatteka@vmware.com>
Signed-off-by: Ansis Atteka <aatteka@vmware.com>