datapath: datapath: Add Modules.supported
This file is required to add external support for SLES 12 SP3 kernel
modules built using the OOT drivers.  Without this file it is necessary
to edit the /etc/modprobe.d/10-unsupported-modules.conf configuration
file to allow load of the Openvswitch OOT kernel drivers.  That is a
security risk because it allows load of any unsupported kernel drivers.
This patch obviates that risk.