rhel/ubuntu/suse: Add "--no-self-confinement" for NSX.
NSX currently (and for a while), adds the "--no-self-confinement"
flag to OVS config files via nsx-agent startup script and
then restarts OVS to pick that config file option.  This flag is needed
to be passed to OVS daemons because we want to allow OVS daemons
to connect to sockets residing outside /var/run/openvswitch.

With systemd, by default, when nsx-agent starts sub-processes,
it adds those sub-processes (in this case OVS) into a common
cgroup. Now, when nsx-agent is stopped, it automatically
kills OVS processes too as the default systemd "killmode" for
processes is "control-group".  nsx-agent cannot control this
as their startup script is sysV. We can explore solutions on
this front, but it may be a better idea to get rid of this
dependency with another component of NSX.  This also prevents
an extra restart of OVS.

With this patch, we do the same job as what nsx-agent does
directly in OVS avoiding an extra step and complications
around process ownership.

Currently nsx-agent would restart ovs if this --no-self-confinement
flag is not present in ovs configuration file. However, if we install
OVS with this flag already set then we can alter execution flow in
nsx-agent so that it would not restart OVS.

This will have to be maintained in the internal OVS repo.

VMware-BZ: #2379002
Signed-off-by: Gurucharan Shetty <guru@ovn.org>