rhel/suse: Start openvswitch after fresh install.
This brings back internal commit in firestar branch -
2f02faf979903 (sles12: enable/start openvswitch with
systemctl post-install). But makes it more correct
by doing it only for installs.  This message also explains
in more detail why we need it.

Current situation is that after RPM install of OVS, OVS daemons
are not automatically started.

In NSX, install of KVM rpms can be done with two workflows.
1. Via mgmt console.
 In this case, there is proper logic to start OVS after installing
OVS RPM.
2. Manually.
 In this case, the customer downloads tar.gz file with all NSX
components and then installs them via "rpm -ivh *.rpm". We have
documented in NSX install guide that the customer should first
install OVS and then "start" OVS and then install other components.
But it is easy to miss.  QA seems to also do this.

Till now, we did not have any issue with this because nsx-agent
startup script would also start OVS forcefully, if not already
started. But, we made changes recently to prevent nsx-agent from
doing it.

For upgrades, it is done via mgmt workflow and they take care of
'force-reload-kmod'. So this patch does not do anything about it.

Upstreaming: One trickiness is that, we are only "starting" on
installs. What about upgrades then? Do we then check kernel versions
and do "force-reload-kmod" too?

VMware-BZ: #2397319
Signed-off-by: Gurucharan Shetty <guru@ovn.org>