Revert "compat: Fix build issue on RHEL 7.7."
This reverts commit f365b41f9dff ("compat: Fix build issue on RHEL 7.7.")
That commit fixed up some issues on RHEL 7.7 and in doing so dropped
support for RHEL kernels earlier than RHEL 7.2 for the sake of cleaning
up the code and making it more readable.  However, we are now asked to
support the 7.0 and 7.1 kernels so revert and create another patch
that will support all RHEL 7.x kernels.

Signed-off-by: Greg Rose <gvrose8192@gmail.com>